+++ sgl_eval/pipeline/__init__.py
@@ -4,6 +4,6 @@ Stage 2 ``spec.run`` : live sampling + scoring + aggregation -> RunResult
 Stage 3 ``report``   : stdout summary + metrics.json + footer + exit code
 """
 
-from sgl_eval.runtime.run import cmd_run
+from sgl_eval.pipeline.run import cmd_run
 
 __all__ = ["cmd_run"]
